+++ src/bin/languageserver/mod.rs
@@ -571,7 +571,7 @@ impl SolangServer {
                 SolangServer::construct_expr(expr3, lookup_tbl, symtab, fnc_map, ns);
             }
 
-            Expression::ArraySubscript(_locs, _typ, expr1, expr2) => {
+            Expression::Subscript(_locs, _typ, expr1, expr2) => {
                 SolangServer::construct_expr(expr1, lookup_tbl, symtab, fnc_map, ns);
                 SolangServer::construct_expr(expr2, lookup_tbl, symtab, fnc_map, ns);
             }