+++ tests/solana_tests/mod.rs
@@ -1,4 +1,5 @@
 mod arrays;
+mod mappings;
 mod primitives;
 mod simple;
 mod storage;